Add files via upload
This is the code that resulted in the final model used for LLNL's DSC. Most of the endgame coding (and training) was done on fxModules.py, predictor.py, and trainer.py.  The results might be improved by training on better-cleaned data and for more epochs.   heartMap.ipynb (soon to be renamed 'heartMachine.ipynb') will need to be fixed to adopt the code from the .py files; this should enable training on a gpu serve.